requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
astunparse==1.6.3
attrs==24.2.0
Automat==24.8.1
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
boto3==1.35.92
botocore==1.35.92
cachetools==5.5.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
cssselect==1.2.0
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.5
fastapi-jwt-auth==0.5.0
fastapi-pagination==0.12.34
filelock==3.16.1
Flask==3.1.0
Flask-PyMongo==3.0.1
flatbuffers==24.12.23
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.4.0
google-auth==2.37.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.1.1
groq==0.11.0
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httptools>=0.6.3,<0.7.0
httpx==0.28.1
huggingface-hub==0.25.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
keras==2.13.1
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.7
langchain-text-splitters==0.3.0
langsmith==0.1.129
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.22.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.24.3
nvidia-cuda-runtime-cu12==12.6.77
oauthlib==3.2.2
openai==1.65.4
opt_einsum==3.4.0
orjson==3.10.7
packaging==24.2
parsel==1.10.0
passlib==1.7.4
pillow==11.1.0
Protego==0.4.0
protobuf==4.25.5
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic>=2.7.4,<3.0.0
pydantic-settings==2.5.2
pydantic_core==2.27.2
PyDispatcher==2.0.7
Pygments==2.18.0
PyJWT==1.7.1
pymongo==4.11.2
PyMuPDF==1.25.4
pyOpenSSL==25.0.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-jose==3.4.0
python-multipart==0.0.20
python-socketio==5.12.1
PyYAML==6.0.2
queuelib==1.7.0
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rich==13.9.1
rsa==4.9
s3transfer==0.10.4
safetensors==0.4.5
schedule==1.2.2
service-identity==24.2.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
tenacity==8.5.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
tldextract==5.1.3
tokenizers==0.20.0
tqdm==4.67.1
tweepy==4.15.0
Twisted==24.11.0
typeguard==4.4.2
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
uvicorn==0.34.0
w3lib==2.2.1
watchfiles==0.24.0
websockets==13.1
Werkzeug==3.1.3
wrapt==1.17.0
wsproto==1.2.0
yarl==1.13.1
zope.interface==7.2
